add reaction updates on comments
